Prognostic run: radiation port available in version of image (#2007)
Packages the python radiation port in `external/radiation` and installs it prognostic run image.

Added public API:
- creates a package `radiation` that can be pip installed from `external/radiation`
- makes it available in the  `prognostic_run_radiation` image

Significant internal changes:
- Makes the radiation port a package with a `setup.py`, and separates the validation test from the radiation port code.
