+++ Makefile
@@ -224,6 +224,7 @@ REQUIREMENTS = external/vcm/setup.py \
 	external/fv3kube/setup.py \
 	external/fv3fit/setup.py \
 	external/loaders/setup.py \
+    external/radiation/setup.py \
 	external/*.requirements.in \
 	workflows/post_process_run/requirements.txt \
 	workflows/prognostic_c48_run/requirements.in \
@@ -257,6 +258,7 @@ docker/prognostic_run/requirements.txt:
 		external/fv3gfs-fortran.requirements.in \
 		external/fv3kube/setup.py \
 		external/vcm/setup.py \
+        external/radiation/setup.py \
 		workflows/post_process_run/requirements.txt \
 		workflows/prognostic_c48_run/requirements.in
