+++ constraints.txt
@@ -2,7 +2,7 @@
 # This file is autogenerated by pip-compile
 # To update, run:
 #
-#    pip-compile --no-annotate --output-file=constraints.txt .dataflow-versions.txt external/fv3fit/setup.py external/fv3gfs-fortran.requirements.in external/fv3kube/setup.py external/loaders/setup.py external/vcm/setup.py pip-requirements.txt projects/microphysics/requirements.in workflows/post_process_run/requirements.txt workflows/prognostic_c48_run/requirements.in
+#    pip-compile --no-annotate --output-file=constraints.txt .dataflow-versions.txt external/fv3fit/setup.py external/fv3gfs-fortran.requirements.in external/fv3kube/setup.py external/loaders/setup.py external/radiation/setup.py external/vcm/setup.py pip-requirements.txt projects/microphysics/requirements.in workflows/post_process_run/requirements.txt workflows/prognostic_c48_run/requirements.in
 #
 absl-py==1.0.0
 aiohttp==3.7.3
@@ -11,7 +11,6 @@ altair==4.1.0
 anyio==2.2.0
 apache-beam==2.37.0
 appdirs==1.4.4
-appnope==0.1.3
 argcomplete==1.12.0
 argon2-cffi==20.1.0
 asciitree==0.3.3
@@ -145,7 +144,7 @@ kubernetes==11.0.0
 lark==1.1.1
 libclang==13.0.0
 libcst==0.4.1
-llvmlite==0.36.0
+llvmlite==0.39.0
 locket==0.2.0
 m2r2==0.3.3
 markdown==3.3.6
@@ -174,7 +173,7 @@ nltk==3.6.7
 nodeenv==1.6.0
 nose==1.3.7
 notebook==6.2.0
-numba==0.53.1
+numba==0.56.0
 numcodecs==0.7.2
 numpy==1.21.5
 oauth2client==4.1.3
@@ -303,6 +302,7 @@ urllib3==1.26.8
 validators==0.18.2
 virtualenv==20.0.33
 wandb==0.12.1
+watchdog==2.1.9
 wcwidth==0.2.5
 webencodings==0.5.1
 websocket-client==1.2.3
