+++ external/radiation/get_data.sh
@@ -2,70 +2,53 @@
 
 set -e
 
-mkdir -p fortran  
+mkdir -p data  
 if [ ! -z "$IS_DOCKER" ] ; then
     echo "This script cannot be run in the Docker image"
 else
 
     export MYHOME=`pwd`
 
-    if [ ! -d "./fortran/data" ]; then
-        mkdir -p fortran
-        cd ./fortran
-        mkdir -p data
-        cd data
-        cd $MYHOME
-    else
-        echo "Fortran output directory already exists"
-    fi
-
-    cd ./python
-    mkdir -p lookupdata
-    cd $MYHOME
-
-    if [ -z "$(ls -A ./python/lookupdata)" ]; then
-        gsutil cp -r gs://vcm-fv3gfs-serialized-regression-data/physics/lookupdata/lookup.tar.gz ./python/lookupdata/.
-        cd ./python/lookupdata
-        tar -xzvf lookup.tar.gz
-        cd $MYHOME
-    else
-        echo "Data already present"
-    fi
-
-    if [ -z "$(ls -A ./fortran/data/radiation_driver)" ]; then
-        gsutil -m cp -r gs://vcm-fv3gfs-serialized-regression-data/physics/radiation_driver/ ./fortran/data/.
-        cd ./fortran/data/radiation_driver
+    if [ ! -d "./data/fortran/radiation_driver)" ]; then
+        mkdir -p ./data/fortran/radiation_driver
+        gsutil -m cp -r gs://vcm-fv3gfs-serialized-regression-data/physics/radiation_driver/* ./data/fortran/radiation_driver
+        cd ./data/fortran/radiation_driver
         tar -xzvf dat_files.tar.gz
         cd $MYHOME
     else
         echo "Driver standalone data already present"
     fi
 
-    if [ ! -d "./python/forcing" ]; then
-        cd ./python
-        mkdir -p forcing
+    if [ ! -d "./data/lookupdata" ]; then
+        cd ./data
+        mkdir lookupdata
+        cd ./lookupdata
+        gsutil cp -r gs://vcm-fv3gfs-serialized-regression-data/physics/lookupdata/lookup.tar.gz .
+        tar -xzvf lookup.tar.gz
         cd $MYHOME
     else
-        echo "Forcing directory already exists"
+        echo "Lookup data already present"
     fi
-
-    if [ -z "$(ls -A ./python/forcing)" ]; then
-	    gsutil -m cp gs://vcm-fv3gfs-serialized-regression-data/physics/forcing/* ./python/forcing/.
-	    cd ./python/forcing
+    
+    if [ ! -d "./data/forcing" ]; then
+        cd ./data
+        mkdir -p forcing
+        cd ./forcing
+	    gsutil -m cp gs://vcm-fv3gfs-serialized-regression-data/physics/forcing/* .
 	    tar -xzvf data.tar.gz
         cd $MYHOME
     else
-	    echo "Forcing data already present"
-    fi  
+        echo "Forcing data already present"
+    fi
 
     if [ "$USE_DIFFERENT_TEST_CASE" != "" ]; then
       echo "Replacing input data with a different namelist"
-      rm -rf ./fortran/data/radiation_driver/*
-      gsutil -m cp -r gs://vcm-fv3gfs-serialized-regression-data/physics/ML_config/input_data_c12_npz63_sw_lw/* ./fortran/data/radiation_driver/.
-      cd ./fortran/data/radiation_driver
+      rm -rf ./data/fortran/radiation_driver/*
+      gsutil -m cp -r gs://vcm-fv3gfs-serialized-regression-data/physics/ML_config/input_data_c12_npz63_sw_lw/* ./data/fortran/radiation_driver/.
+      cd ./data/fortran/radiation_driver
       tar -xvf dat_files.tar.gz
       cd $MYHOME
-      cd ./python/lookupdata
+      cd ./data/lookupdata
       rm rand2d_tile*.nc
       gsutil -m cp -r gs://vcm-fv3gfs-serialized-regression-data/physics/ML_config/random_MLconfig/* .
     fi
