+++ external/radiation/radiation/radsw/__init__.py
@@ -1,0 +1,1 @@
+from .radsw_main import RadSWClass
