+++ external/radiation/radiation/types.py
@@ -1,0 +1,3 @@
+import numpy as np
+
+DTYPE_INT = np.int32
