+++ workflows/prognostic_c48_run/tests/test_radiation.py
@@ -1,0 +1,3 @@
+def test_radiation_import():
+    """Just imports the driver, could be extended"""
+    from radiation import RadiationDriver  # noqa: F401
